+++ Surfacer/src/core/ViewportController.hpp
@@ -40,14 +40,14 @@ namespace core {
 
 		 */
 		struct zeno_config {
-			float panFactor, scaleFactor;
+			double panFactor, scaleFactor;
 
 			zeno_config():
 			panFactor(0.01),
 			scaleFactor(0.01)
 			{}
 
-			zeno_config( float pf, float zf ):
+			zeno_config( double pf, double zf ):
 			panFactor(pf),
 			scaleFactor(zf)
 			{}
@@ -78,27 +78,27 @@ namespace core {
 		/**
 			Set the scale
 		 */
-		void setScale( float z ) { _scale = _constrainScale(z); }
+		void setScale( double z ) { _scale = _constrainScale(z); }
 
 		/**
 			set the scale and pan such that @a about stays in the same spot on screen
 		 */
-		void setScale( float z, const vec2 &about );
+		void setScale( double z, const dvec2 &about );
 
 		/**
 			get the current scale
 		 */
-		float getScale() const { return _scale; }
+		double getScale() const { return _scale; }
 
 		/**
 			set the pan
 		 */
-		void setPan( const vec2 &p ) { _pan = _constrainPan( p ); }
+		void setPan( const dvec2 &p ) { _pan = _constrainPan( p ); }
 
 		/**
 			get the current pan
 		 */
-		vec2 getPan() const {
+		dvec2 getPan() const {
 			return _pan;
 		}
 
@@ -111,23 +111,23 @@ namespace core {
 			This is the function to use if, for example, you want the camera to pan to some
 			object or character of interest.
 		 */
-		void lookAt( const vec2 &world, float scale, const vec2 &screen );
+		void lookAt( const dvec2 &world, double scale, const dvec2 &screen );
 
 		/**
-			Shortcut to lookAt( const vec2 &world, float scale, const vec2 &screen )
+			Shortcut to lookAt( const dvec2 &world, double scale, const dvec2 &screen )
 		 */
-		void lookAt( const vec2 &world, float scale ) { lookAt( world, scale, _viewport.getViewportCenter() ); }
+		void lookAt( const dvec2 &world, double scale ) { lookAt( world, scale, _viewport.getViewportCenter() ); }
 
 		/**
-			Shortcut to lookAt( const vec2 &world, float scale, const vec2 &screen )
+			Shortcut to lookAt( const dvec2 &world, double scale, const dvec2 &screen )
 			Causes the camera to center @a world in the screen
 		 */
-		void lookAt( const vec2 &world ) { lookAt( world, _scale, _viewport.getViewportCenter() ); }
+		void lookAt( const dvec2 &world ) { lookAt( world, _scale, _viewport.getViewportCenter() ); }
 
 	protected:
 
-		vec2 _constrainPan( vec2 p ) const;
-		float _constrainScale( float z ) const;
+		dvec2 _constrainPan( dvec2 p ) const;
+		double _constrainScale( double z ) const;
 
 		virtual void _stepZeno( const core::time_state &time );
 		virtual void _stepPID( const core::time_state &time );
@@ -141,8 +141,8 @@ namespace core {
 		unsigned int		_constraintMask;
 		control_method		_controlMethod;
 		cpBB				_levelBounds;
-		float				_scale;
-		vec2				_pan;
+		double				_scale;
+		dvec2				_pan;
 		zeno_config			_zenoConfig;
 		bool				_disregardViewportMotion;
 	};
